+++ plugin/src/test/kotlin/dev/nesk/akkurate/ksp/ValidateAnnotationProcessorTest.kt
@@ -248,6 +248,32 @@ class ValidateAnnotationProcessorTest {
     }
 
     @Test
+    fun `accessors are generated only for public properties`() {
+        // Arrange
+        val source = SourceFile.kotlin(
+            "Examples.kt", """
+                package dev.nesk
+                import dev.nesk.akkurate.annotations.Validate
+                private class PrivateParent {
+                    public class PublicParent {
+                        @Validate public class PublicClass(public val publicPropInPrivateScope: Any)
+                    }
+                }
+                @Validate public class PublicClass(public val publicPropInPublicScope: Any, private val privatePropInPublicScope: Any)
+            """
+        )
+
+        // Act
+        val (result, compiler) = compile(source)
+
+        // Assert
+        assertCompilationIsSuccessful(result)
+        assertCountOfFilesGeneratedByTheProcessor(1, compiler)
+        Path("${compiler.kotlinFilesDir}/dev/nesk/validation/accessors/ValidationAccessors.kt").readText()
+            .matchWithSnapshot("accessors are generated only for public properties")
+    }
+
+    @Test
     fun `the option 'appendPackagesWith' appends its value to the original package name`() {
         // Arrange
         val source = SourceFile.kotlin(